requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.15.0
pyyaml>=6.0.1
pandas>=2.0.0
numpy>=1.24.0
wandb>=0.15.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
trl>=0.7.4